# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "client_message_id": {
            "format": "uuid",
            "title": "Client Message Id",
            "type": "string"
          },
          "message": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": ["message", "client_message_id"],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "ChatSession": {
        "additionalProperties": false,
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ChatSessionStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": ["id", "title", "status", "created_at", "updated_at"],
        "title": "ChatSession",
        "type": "object"
      },
      "ChatSessionCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "default": "New chat",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "title": "ChatSessionCreateRequest",
        "type": "object"
      },
      "ChatSessionDetail": {
        "additionalProperties": false,
        "properties": {
          "active_turn": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatTurn"
              },
              {
                "type": "null"
              }
            ]
          },
          "messages": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "session": {
            "$ref": "#/components/schemas/ChatSession"
          }
        },
        "required": ["session"],
        "title": "ChatSessionDetail",
        "type": "object"
      },
      "ChatSessionPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ChatSession"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "ChatSessionPage",
        "type": "object"
      },
      "ChatSessionStatus": {
        "enum": ["active", "archived"],
        "title": "ChatSessionStatus",
        "type": "string"
      },
      "ChatSessionUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ChatSessionUpdateRequest",
        "type": "object"
      },
...
      "ChatTurnAccepted": {
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ChatTurnStatus"
          },
          "stream_url": {
            "title": "Stream Url",
            "type": "string"
          },
          "turn_id": {
            "format": "uuid",
            "title": "Turn Id",
            "type": "string"
          },
          "user_message_id": {
            "format": "uuid",
            "title": "User Message Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "turn_id",
          "user_message_id",
          "status",
          "stream_url"
        ],
        "title": "ChatTurnAccepted",
        "type": "object"
      },
...
      "CommandRequest": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Command",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          }
        },
        "required": ["command"],
        "title": "CommandRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunSourcesResponse": {
        "additionalProperties": false,
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sources": {
            "default": [],
            "items": {},
            "title": "Sources",
            "type": "array"
          }
        },
        "required": ["run_id"],
        "title": "RunSourcesResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Domain-agnostic capability platform foundation.",
    "title": "TradeSentinel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "operationId": "chat_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTurnAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat",
        "tags": ["chat"]
      }
    },
    "/api/v1/chat/sessions": {
      "get": {
        "operationId": "chat_sessions_api_v1_chat_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Sessions",
        "tags": ["chat"]
      },
      "post": {
        "operationId": "create_chat_session_api_v1_chat_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Chat Session",
        "tags": ["chat"]
      }
    },
    "/api/v1/chat/sessions/{session_id}": {
      "get": {
        "operationId": "chat_session_api_v1_chat_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Session",
        "tags": ["chat"]
      },
      "patch": {
        "operationId": "update_chat_session_api_v1_chat_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Chat Session",
        "tags": ["chat"]
      }
    },
    "/api/v1/chat/turns/{turn_id}": {
      "get": {
        "operationId": "get_chat_turn_api_v1_chat_turns__turn_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTurn"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chat Turn",
        "tags": ["chat"]
      }
    },
    "/api/v1/chat/turns/{turn_id}/events": {
      "get": {
        "operationId": "chat_turn_events_api_v1_chat_turns__turn_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "turn_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Turn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Turn Events",
        "tags": ["chat"]
      }
    },
...
    "/api/v1/commands/execute": {
      "post": {
        "operationId": "execute_command_api_v1_commands_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Command",
        "tags": ["commands"]
      }
    },
...
    "/api/v1/instruments/{symbol}/history": {
      "get": {
        "operationId": "instrument_history_api_v1_instruments__symbol__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Instrument History",
        "tags": ["future"]
      }
    },
    "/api/v1/instruments/{symbol}/quote": {
      "get": {
        "operationId": "instrument_quote_api_v1_instruments__symbol__quote_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Instrument Quote",
        "tags": ["future"]
      }
    },
...
    "/api/v1/predictions/{prediction_id}": {
      "get": {
        "operationId": "get_prediction_api_v1_predictions__prediction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prediction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prediction",
        "tags": ["future"]
      }
    },
    "/api/v1/predictions/{prediction_id}/evaluation": {
      "get": {
        "operationId": "get_prediction_evaluation_api_v1_predictions__prediction_id__evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prediction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prediction Evaluation",
        "tags": ["future"]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Run Api V1 Runs  Run Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": ["runs"]
      }
    },
    "/api/v1/runs/{run_id}/sources": {
      "get": {
        "operationId": "get_run_sources_api_v1_runs__run_id__sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSourcesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Sources",
        "tags": ["runs"]
      }
    },
    "/api/v1/workflows/{workflow_name}/execute": {
      "post": {
        "operationId": "execute_workflow_api_v1_workflows__workflow_name__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_name",
            "required": true,
            "schema": {
              "title": "Workflow Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Execute Workflow",
        "tags": ["workflows"]
      }
    },
...
  }
}